refactor: extract ResourceTableMenu component
